chore: provide platform-specific web video wrappers
